adding standardize tests
